Describe a complex survey design

Returns a one-row tibble of design diagnostics: sample size \(n\), number of strata \(H\), total number of PSUs \(\sum_h \alpha_h\), and summary statistics of the adjusted weights \(w_k\) including their coefficient of variation \(CV(w) = s_w / \bar{w}\), which is an indicator of weight variability and its effect on variance inflation.

describe_survey_design(design)

Arguments

design

A tbl_svy object created with as_survey_design_tbl.

Value

A tibble with one row and the following columns:

n_obs

Total sample size \(n = |s|\).

n_strata

Number of strata \(H\) (NA if no stratification variable was supplied).

n_clusters

Total number of PSUs \(\sum_{h=1}^{H} \alpha_h\) (NA if no cluster variable was supplied).

weight_min

\(\min_{k \in s} w_k\).

weight_max

\(\max_{k \in s} w_k\).

weight_mean

Unweighted mean of the adjusted weights \(\bar{w} = \hat{N}_w / n\), where \(\hat{N}_w = \sum_{k \in s} w_k\).

weight_cv

Coefficient of variation of the weights \(CV(w) = s_w / \bar{w}\). Large values (\(> 0.5\)) indicate high weight variability, which can substantially inflate the design effect \(\widehat{\mathrm{DEFF}}\).

See also

Examples

data <- generate_example_data(30)
design <- as_survey_design_tbl(
  data,
  weight  = "weight",
  strata  = "strata",
  cluster = "upm"
)
describe_survey_design(design)
#> # A tibble: 1 × 7
#>   n_obs n_strata n_clusters weight_min weight_max weight_mean weight_cv
#>   <int>    <int>      <int>      <dbl>      <dbl>       <dbl>     <dbl>
#> 1   997        5         30       202.      1493.        864.     0.443